ffffuuuu

+++ src/Peer.java
@@ -358,7 +358,7 @@ public class Peer extends java.rmi.server.UnicastRemoteObject implements PeerInt
 		
 		FileElement localFile = localList.get(localList.indexOf(file));
 		//Local file dirty bit not set, so can just replace the local file.
-		if (localFile.changed == false)
+		if ((localFile.changed == false) && (file.changed == true))
 		{
 			//Remove the file from the filesystem
 			boolean delsuccess = f.delete();
@@ -404,7 +404,6 @@ public class Peer extends java.rmi.server.UnicastRemoteObject implements PeerInt
 			fe.changed = true;
 			notifyPeersChanged(fe);
 		}
-		
 	}
 	
 	private int removeLocalFile(FileElement file){
